Add `COMPOSER_AUTH_JSON_LOCATION` support (#5)
Pass `COMPOSER_AUTH_JSON_LOCATION` env var 

Allows you to define the location of a `auth.json` composer auth file to be mounted into the container for use with custom repositories etc.

Also mount in the assets instead of copying them into the docker image. We need all the associated infra of this repo to be able to use the image in the manner we desire in any case. So we can mount in the dockerfile-assets directory so that we can more easily make changes to and debug the scripts